+++ src/menuIO/xmlFmt.h
@@ -6,48 +6,29 @@
 
   namespace Menu {
 
-    Print& operator<<(Print&o, Menu::prompt&p) {
+    inline Print& operator<<(Print&o, Menu::prompt&p) {
       print_P(o,p.getText());
       return o;
     }
 
-    menuOut& operator<<(menuOut&o, idx_t i) {
+    inline menuOut& operator<<(menuOut&o, idx_t i) {
       o.print(i);
       return o;
     }
 
-    menuOut& operator<<(menuOut&o, const char* p) {
+    inline menuOut& operator<<(menuOut&o, const char* p) {
       print_P(o,p);
       return o;
     }
 
-    Print& operator<<(Print&o, const char* p) {
+    inline Print& operator<<(Print&o, const char* p) {
       print_P(o,p);
       return o;
     }
 
-    menuOut& operator<<(menuOut&o,classes c) {
-      switch(c) {
-        case noClass:return o<<"noClass";break;
-        case promptClass:return o<<"prompt";break;
-        case fieldClass:return o<<"field";break;
-        case toggleClass:return o<<"toggle";break;
-        case selectClass:return o<<"select";break;
-        case chooseClass:return o<<"choose";break;
-        case valueClass:return o<<"value";break;
-        case menuClass:return o<<"menu";break;
-        default: return o;
-      }
-    }
+    menuOut& operator<<(menuOut&o,classes c);
 
-    void outputOptions(menuOut& o,navNode &nav,menuNode& node,idx_t idx) {
-      o<<"<node data-path=\"/menu?at=";
-      nav.root->printPath(o);
-      o<<"/"<<idx<<"\">";
-      for(idx_t n=0;n<node.sz();n++)
-        o<<"<value"<<(n==node.selected()?" selected=\"selected\"":"")<<">"<<node[n]<<"</value>";
-      o<<"</node>";
-    }
+    void outputOptions(menuOut& o,navNode &nav,menuNode& node,idx_t idx);
 
     template<class T>
     class xmlFmt:public T {
